fixes #19
The "agents_on" attributes for road is not used anymore in the road skill, so I removed it. You can use "all_agents" to get the list of agents on the road. I removed as well the "unregister" in the road skill that was deprecated (now, unregister has to be used in the driving skill).