Commit on the runtime of simulations and synchronizations with displays
- all the management done with ExecutorService (for simulations) and
volatile boolean variables (for the rendition of displays) is replaced
by (simpler) semaphores;
- The acquisition of the semaphores now occur at more relevant places
(allowing to synchronize also the 2D/3D renderer threads for a more
accurate synchronization -- see #124).
- Each simulation now keeps its own thread throughout the experiment:
easier for debugging, and also makes it possible to define ThreadLocal
variables in shared objects (in skills, for instance) that will remain
attached to the simulation

*** LOTS OF TESTS ARE NECESSARY TO ENSURE THAT IT CAN BE A GOOD
FOUNDATION FOR THE GAMA RUNTIME ***

Known issues:
- no time out is set (yet) for the acquisition of the semaphore, which
can potentially lead to infinite loop/wait when the rendition is not
done properly
- the max. number of threads to use for the execution of simulations is
not (yet) respected
- errors are visible in the console when exiting GAMA and a simulation
is running.
- the release of signals from the semaphores is not fair (yet).
- the problem where a chart in 2d slows down all the other displays when
the mouse or keyboard events are targeted to it is not (yet) solved.