Fixes #622
Adds a flag in GamaPopulation to know when it is disposed of. Based on this flag, let the individual agents remove themselves or not from their population + topology. In case of disposal, the topologies are being taken care of by the simulation, so no additional need to dispose them in each population. Fixes completely #622 and even seems to improve the memory at the end of an experiment (less temporary "left overs" like Envelope3D objects, etc.)