Improve class handling in GAML model factories
This commit refactors and extends support for 'class' statements in GAML. It updates factories to properly create and register ClassDescription instances, adjusts type checks and facet types for classes, and fixes a typo in Types.java. These changes improve the extensibility and correctness of class definitions and inheritance in GAML models.